refactor: 도메인별 테스트 폴더 생성
