+++ src/main/java/gift/domain/user/dto/UserDTO.java
@@ -1,7 +1,5 @@
 package gift.domain.user.dto;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import gift.domain.user.User;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
