+++ src/test/java/gift/CartItem/CartItemRepositoryTest.java
@@ -1,4 +1,4 @@
-package gift.JpaDataTest;
+package gift.CartItem;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
