+++ src/test/java/gift/Category/CategoryRepositoryTest.java
@@ -1,4 +1,4 @@
-package gift.JpaDataTest;
+package gift.Category;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
