+++ src/test/java/gift/Option/OptionRepositoryTest.java
@@ -1,4 +1,4 @@
-package gift.JpaDataTest;
+package gift.Option;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
