+++ src/test/java/gift/Product/ProductRepositoryTest.java
@@ -1,4 +1,4 @@
-package gift.JpaDataTest;
+package gift.Product;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
